\name{meteogRam-internal}
\alias{humi}
\alias{temperature.data}
\alias{temperature}
\alias{wind}
\title{Internal meteogRam Functions}
\description{
 Internal meteogRam functions
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}
\keyword{ internal }
